+++ lifemonitor/commands/tasks.py
@@ -40,7 +40,7 @@ def reset():
     Reset task-queue status
     """
     from lifemonitor.cache import clear_cache
-    from lifemonitor.tasks.task_queue import REDIS_NAMESPACE
+    from lifemonitor.tasks.config import REDIS_NAMESPACE
     try:
         clear_cache(client_scope=False, prefix=REDIS_NAMESPACE)
     except Exception as e:
